Move command summary files
