+++ artifactory/utils/commandsummary/uploadsummary.go
@@ -1,9 +1,8 @@
-package generic
+package commandsummary
 
 import (
 	"fmt"
 	"github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
-	"github.com/jfrog/jfrog-cli-core/v2/artifactory/utils/commandsummary"
 )
 
 type UploadSummary struct {
@@ -23,8 +22,8 @@ type ResultsWrapper struct {
 	Results []UploadResult `json:"results"`
 }
 
-func NewUploadSummary(platformUrl string, majorVersion int) (*commandsummary.CommandSummary, error) {
-	return commandsummary.New(&UploadSummary{
+func NewUploadSummary(platformUrl string, majorVersion int) (*CommandSummary, error) {
+	return New(&UploadSummary{
 		platformUrl:  platformUrl,
 		majorVersion: majorVersion,
 	}, "upload")
@@ -44,7 +43,7 @@ func (us *UploadSummary) loadResults(filePaths []string) error {
 	us.uploadedArtifacts = ResultsWrapper{}
 	for _, path := range filePaths {
 		var uploadResult ResultsWrapper
-		if err := commandsummary.UnmarshalFromFilePath(path, &uploadResult); err != nil {
+		if err := UnmarshalFromFilePath(path, &uploadResult); err != nil {
 			return err
 		}
 		us.uploadedArtifacts.Results = append(us.uploadedArtifacts.Results, uploadResult.Results...)
@@ -61,5 +60,5 @@ func (us *UploadSummary) generateFileTreeMarkdown() string {
 }
 
 func (us *UploadSummary) buildUiUrl(targetPath string) string {
-	return commandsummary.GenerateArtifactUrl(us.platformUrl, targetPath, us.majorVersion)
+	return GenerateArtifactUrl(us.platformUrl, targetPath, us.majorVersion)
 }
