+++ artifactory/utils/testdata/command_summaries/image2.md
