+++ artifactory/utils/testdata/command_summaries/multiarch-image1.md
